+++ card-capture/src/components/editor/EditingArea/components/FocusBox/FocusBox.tsx
@@ -3,12 +3,13 @@
 import { useCardsStore } from '@/store/useCardsStore';
 import { LayerType, Position } from '@/store/useCardsStore/type';
 import React, { useEffect, useRef, useState } from 'react';
-import { Direction, Offset, ResizeOffset } from './FocusBox.type';
+import { Direction, ResizeOffset } from './FocusBox.type';
 import { FaArrowRotateLeft } from 'react-icons/fa6';
-import { INITIAL_DRAG_OFFSET, INITIAL_RESIZE_OFFSET } from './FocusBox.constant';
+import { INITIAL_RESIZE_OFFSET } from './FocusBox.constant';
 import { useFocusStore } from '@/store/useFocusStore';
 import { ContextMenu, ContextMenuContent, ContextMenuItem, ContextMenuTrigger } from '@/components/ui/context-menu';
 import TrashIcon from '@/components/common/Icon/TrashIcon';
+import useDrag from '@/components/editor/EditingArea/components/FocusBox/hooks/useDrag';
 
 type Props = {
   children: React.ReactElement<{
@@ -29,13 +30,7 @@ const FocusBox = ({ children, cardId, layerId, type, initialMouseDown }: Props) 
   const layer = useCardsStore(state => state.cards[cardId].layers.filter(v => v.id === layerId)[0]);
   const setPosition = useCardsStore(state => state.setPosition);
 
-  const [curPosition, setCurPosition] = useState(layer.position); //현재 위치를 스토어에 업로드 하지 않고 관리하기위한 state
-
-  //drag State
-  const [isDrag, setIsDrag] = useState(false);
-  const [dragOffset, setDragOffset] = useState<Offset>({
-    ...INITIAL_DRAG_OFFSET,
-  });
+  const [curPosition, setCurPosition] = useState(layer.position); // 현재 위치를 스토어에 업로드 하지 않고 관리하기위한 state
 
   //resize State
   const [resizeState, setResizeState] = useState<Direction>('none');
@@ -74,80 +69,15 @@ const FocusBox = ({ children, cardId, layerId, type, initialMouseDown }: Props) 
   /* 드래그 관련 로직 */
   //              //
 
-  /**
-   * layer -> focus로 변경될 때 mouseDown 이벤트 전달해서 바로 드래그 되도록 하는 로직
-   * @TODO : 추추에는 layerbox를 없애고 focusBox로만 관리하는 방식으로 변경할 예정
-   */
-
-  useEffect(() => {
-    // initialMouseDown이 있으면 즉시 드래그 시작
-    if (initialMouseDown) {
-      const startDrag = (e: React.MouseEvent) => {
-        setIsDrag(true);
-        setDragOffset({
-          x: e.clientX - curPosition.x,
-          y: e.clientY - curPosition.y,
-        });
-      };
-      startDrag(initialMouseDown);
-    }
-  }, [initialMouseDown]);
-
-  const pointerDownDragHandler = (e: React.PointerEvent) => {
-    e.stopPropagation();
-    setIsDrag(true);
-
-    setDragOffset(prev => {
-      return {
-        ...prev,
-        y: e.clientY - layer.position.y,
-        x: e.clientX - layer.position.x,
-      };
-    });
-  };
-
-  const pointerMoveDragHandler = (e: PointerEvent) => {
-    if (!isDrag) return;
-
-    const diffX = e.clientX - dragOffset.x;
-    const diffY = e.clientY - dragOffset.y;
-
-    setCurPosition(prev => {
-      return {
-        ...prev,
-        x: diffX,
-        y: diffY,
-      };
-    });
-  };
-
-  /**
-   * @NOTE: 이동이 끝날 때 cardStore에 저장
-   */
-  const pointerUpDragHandler = (e: PointerEvent) => {
-    const diffX = e.clientX - dragOffset.x;
-    const diffY = e.clientY - dragOffset.y;
-
-    setDragOffset({ ...INITIAL_DRAG_OFFSET });
-    setIsDrag(false);
-    setPosition(cardId, layerId, { ...curPosition, y: diffY - 0.4, x: diffX - 0.4 });
-  };
-
-  /**
-   * 드래그 이벤트 등록
-   */
-  useEffect(() => {
-    if (!isDrag) return;
-    if (isDoubleClicked && type === 'text') return;
-
-    window.addEventListener('pointermove', pointerMoveDragHandler);
-    window.addEventListener('pointerup', pointerUpDragHandler);
-
-    return () => {
-      window.removeEventListener('pointermove', pointerMoveDragHandler);
-      window.removeEventListener('pointerup', pointerUpDragHandler);
-    };
-  }, [isDrag]);
+  const { isDrag, pointerDownDragHandler } = useDrag({
+    cardId,
+    layerId,
+    type,
+    curPosition,
+    setCurPosition,
+    initialMouseDown,
+    isDoubleClicked,
+  });
 
   //                //
   /* 크기 resize 로직 */
